+++ spec/unit/imap/backup/cli/utils_spec.rb
@@ -1,4 +1,4 @@
-require "imap/backup/serializer/folder_maker"
+require "imap/backup/cli/utils"
 require "support/shared_examples/an_action_that_handles_logger_options"
 
 module Imap::Backup