+++ spec/unit/imap/backup/serializer/unused_name_finder_spec.rb
@@ -1,3 +1,6 @@
+require "imap/backup/serializer/unused_name_finder"
+require "imap/backup/serializer"
+
 module Imap::Backup
   RSpec.describe Serializer::UnusedNameFinder do
     subject { described_class.new(serializer: serializer) }